requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.0.3
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.3
Django==3.2.11
kombu==5.2.3
pip==21.1.2
prompt-toolkit==3.0.24
python-crontab==2.6.0
python-dateutil==2.8.2
python-libnmap==0.7.2
pytz==2021.3
setuptools==59.6.0
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wheel==0.36.2
